pyproject.toml
[project]
name = "github-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=1.40.0",
    "python-dotenv>=1.0.1",
    "streamlit>=1.36.0",
]
